@@ -10,15 +10,17 @@ import (
 type OrderService struct {
 	pb.UnimplementedOrderServiceServer
 	CreateOrderUseCase usecase.CreateOrderUseCase
+	listOrderUseCase   usecase.ListOrderUseCase
 }
 
-func NewOrderService(createOrderUseCase usecase.CreateOrderUseCase) *OrderService {
+func NewOrdersService(createOrderUseCase usecase.CreateOrderUseCase, listOrderUseCase usecase.ListOrderUseCase) *OrderService {
 	return &OrderService{
 		CreateOrderUseCase: createOrderUseCase,
+		listOrderUseCase:   listOrderUseCase,
 	}
 }
 
-func (s *OrderService) CreateOrder(ctx context.Context, in *pb.CreateOrderRequest) (*pb.CreateOrderResponse, error) {
+func (s *OrderService) CreateOrder(ctx context.Context, in *pb.CreateOrderRequest) (*pb.Order, error) {
 	dto := usecase.OrderInputDTO{
 		ID:    in.Id,
 		Price: float64(in.Price),
@@ -28,10 +30,29 @@ func (s *OrderService) CreateOrder(ctx context.Context, in *pb.CreateOrderReques
 	if err != nil {
 		return nil, err
 	}
-	return &pb.CreateOrderResponse{
+	return &pb.Order{
 		Id:         output.ID,
 		Price:      float32(output.Price),
 		Tax:        float32(output.Tax),
 		FinalPrice: float32(output.FinalPrice),
+	}, nil
+}
+
+func (s *OrderService) ListOrders(ctx context.Context, req *pb.Blank) (*pb.OrderList, error) {
+	orders, err := s.listOrderUseCase.Execute()
+	if err != nil {
+		return nil, err
+	}
+	var ordersResponse []*pb.Order
+	for _, order := range orders {
+		ordersResponse = append(ordersResponse, &pb.Order{
+			Id:         order.ID,
+			Price:      float32(order.Price),
+			Tax:        float32(order.Tax),
+			FinalPrice: float32(order.FinalPrice),
+		})
+	}
+	return &pb.OrderList{
+		Orders: ordersResponse,
 	}, nil
 }
